backend/requirements.txt
# ── Web Framework ──────────────────────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Data Validation ─────────────────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
email-validator==2.1.1

# ── Database ────────────────────────────────────────────────────────────────────
supabase==2.4.6
asyncpg==0.29.0

# ── Auth ────────────────────────────────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── RAG / AI ───────────────────────────────────────────────────────────────────
openai==1.30.1
pinecone-client==3.2.2
tiktoken==0.7.0
anthropic==0.26.0

# ── ML ─────────────────────────────────────────────────────────────────────────
scikit-learn==1.4.2
xgboost==2.0.3
imbalanced-learn==0.12.2
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2

# ── NLP ─────────────────────────────────────────────────────────────────────────
spacy==3.7.4

# ── PDF / OCR ──────────────────────────────────────────────────────────────────
pdfplumber==0.11.1
pdf2image==1.17.0
pytesseract==0.3.10
Pillow==10.3.0

# ── HTTP / Async ───────────────────────────────────────────────────────────────
httpx==0.27.0
aiofiles==23.2.1

# ── Utilities ──────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
structlog==24.1.0
tenacity==8.3.0

# ── Testing ────────────────────────────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.6
pytest-cov==5.0.0
httpx==0.27.0
